requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
passlib==1.7.4
psycopg2-binary==2.9.10
python-jose==3.5.0
Jinja2==3.1.6
pydantic[email]
python-multipart